chore(cg): Separate building logic
Closes #243
