+++ src/internals/ir/builders/index.ts
@@ -1,2 +1,4 @@
 export * from "./ir";
 export * from "./astStore";
+export * from "./callgraph";
+export * from "./imports";
